+++ Pod/Core/NSObject+ActorKit.h
@@ -22,7 +22,6 @@ FOUNDATION_EXPORT NSString * const TBAKActorPayload;
  */
 typedef void (^TBActorPoolConfigurationBlock)(NSObject *actor, NSUInteger index);
 
-@protocol TBActorSupervision;
 @class TBActorPool;
 
 /**
@@ -34,11 +33,6 @@ typedef void (^TBActorPoolConfigurationBlock)(NSObject *actor, NSUInteger index)
  *  The actor's operation queue.
  */
 @property (nonatomic, strong) NSOperationQueue *actorQueue;
-
-/**
- *  Reference to the current supervisor. Can be messaged through TBActorSupervison protocol.
- */
-@property (nonatomic, weak) NSObject <TBActorSupervision> *supervisor;
 
 /**
  *  The pool the actor may belong to.
@@ -107,13 +101,6 @@ typedef void (^TBActorPoolConfigurationBlock)(NSObject *actor, NSUInteger index)
  *  @param payload     The payload of the notification.
  */
 - (void)publish:(NSString *)messageName payload:(id)payload;
-
-/**
- *  Notifies the supervisor about the crash using the TBActorSupervison protocol.
- *
- *  @param error The optional error.
- */
-- (void)crashWithError:(NSError *)error;
 
 /**
  *  Creates a pool of actors of the current class using a specified configuration block.